Adjusting the file structure
I have created a main.py file in order to have that be the entry point for the project. Thus, for further expansions which may require additional support files, the functionality of the project will be linked back to main.py

Also,  I have adjusted the README.md in order to include the necessary installations in order to run this project.